Implement smaller refactors with clang-tidy suggestions
- add incomplete parse function that replaces ENV Vars with actual value (faulty)

+++ include/history.h
@@ -3,6 +3,6 @@
 
 void append_to_history(const char *command);
 void print_history();
-void execute_command_from_history(int index);
+void execute_command_from_history(unsigned long index);
 
 #endif
